+++ backend/src/test/java/ku_rum/backend/domain/user/application/MailServiceTest.java
@@ -1,15 +1,15 @@
 package ku_rum.backend.domain.user.application;
 
-import ku_rum.backend.domain.user.application.mail.MailService;
-import ku_rum.backend.domain.user.dto.request.mail.MailSendRequest;
-import ku_rum.backend.domain.user.dto.request.mail.MailVerificationRequest;
+import ku_rum.backend.domain.mail.application.MailService;
+import ku_rum.backend.domain.mail.dto.request.MailSendRequest;
+import ku_rum.backend.domain.mail.dto.request.MailVerificationRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.redis.core.RedisTemplate;
 
-import static ku_rum.backend.domain.user.domain.MailSendSetting.MAIL_SEND_INFO;
+import static ku_rum.backend.domain.mail.domain.MailSendSetting.MAIL_SEND_INFO;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
